update strings 11/23

+++ app/src/main/java/com/example/guanghuahe/cst2335_finalmilestone1/movie/fragments/HistoryToolBarFragment.java
@@ -1,5 +1,6 @@
 package com.example.guanghuahe.cst2335_finalmilestone1.movie.fragments;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -25,10 +26,14 @@ public class HistoryToolBarFragment extends Fragment {
         private Button removeAll, byYear, statistic, byRuntime, byRating;
         private DatabaseHelper DB;
         private HistoryFragment historyFragment;
+        private Context mainActivity;
 
 
-
-
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        mainActivity = context;
+    }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -50,9 +55,18 @@ public class HistoryToolBarFragment extends Fragment {
 
 
         removeAll.setOnClickListener(e-> {
-                    DB.removeAll();
-            historyFragment.updateView();
+            AlertDialog dialog = new AlertDialog.Builder(mainActivity)
+                    .setTitle("DELETE ALL")
+                    .setMessage(R.string.dialog_message)
+                    .setPositiveButton("Sure", (dialog1, which) ->{
+                        DB.removeAll();
+                        historyFragment.updateView();
+                    })
+                    .setNegativeButton(R.string.dialog_button_cancel,(dialog1,which)->{
+                        dialog1.dismiss();
+                    }).create();
 
+                    dialog.show();
                 });
         byRuntime.setOnClickListener(e->{
             historyFragment.orderBy("runtime");
